requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0
torch==2.0.1
transformers==4.30.2

# Data Processing
pyarrow==12.0.1
fastparquet==2023.7.0
joblib==1.3.1

# API and Web
flask==2.3.2
flask-cors==4.0.0
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3

# AWS Integration
boto3==1.28.17
sagemaker==2.172.0

# Monitoring and Logging
mlflow==2.5.0
wandb==0.15.5
prometheus-client==0.17.1

# Privacy and Security
cryptography==41.0.2
hashlib-additional==1.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.65.0
click==8.1.6

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Model Interpretation
shap==0.42.1
lime==0.2.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.4.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.0
